Split chattingMenu into handler and menu
